Improve thread group size support in shader reflection
Update interop signatures and logic to enhance thread group size handling for compute, mesh, and amplification shaders. Update XML docs, switch buffer/code size types to nuint, add new P/Invoke bindings, and introduce GetThreadGroupSizes. Assign native thread group sizes during deserialization for relevant stages, ensuring accurate cross-platform reflection.

+++ Slangc.NET/SlangBlob.cs
@@ -32,6 +32,6 @@ public unsafe struct SlangBlob
     {
         SlangBlob* @this = (SlangBlob*)Unsafe.AsPointer(ref Unsafe.AsRef(in this));
 
-        return ((delegate* unmanaged[Stdcall]<SlangBlob*, ulong>)@this->LpVtbl[4])(@this);
+        return ((delegate* unmanaged[Stdcall]<SlangBlob*, nuint>)@this->LpVtbl[4])(@this);
     }
 };
